Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 void main(){
     float r;
     scanf("%f", &r);
-    int area = 3.14 * r * r;
+    float area = 3.14 * r * r;
     scanf("%f",&area);
     printf("Area: %.2f\n",area);
 }
